Added seeders, pagination, filters

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Category;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CategoryRequest;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
@@ -16,7 +17,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        $categories = Category::get();
+        $categories = Category::paginate(10);
 
         return view('auth.categories.index', compact('categories'));
     }
@@ -37,12 +38,16 @@ class CategoryController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CategoryRequest $request)
     {
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
-        $cat = Category::create($params);
+        unset($params['image']);
+
+        if ($request->has('image')) {
+            $params['image'] = $request->file('image')->store('categories');
+        }
+
+        Category::create($params);
         return redirect()->route('categories.index');
     }
 
@@ -75,12 +80,15 @@ class CategoryController extends Controller
      * @param  \App\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Category $category)
+    public function update(CategoryRequest $request, Category $category)
     {
-        Storage::delete($category->image);
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+
+        if ($request->has('image')) {
+            Storage::delete($category->image);
+            $params['image'] = $request->file('image')->store('categories');
+        }
 
         $category->update($params);
         return redirect()->route('categories.index');
